backend/requirements_fashn_tester.txt
requests>=2.31.0
Pillow>=9.5.0
gradio>=3.35.0
python-dotenv>=1.0.0 